added the delete function and the route,fixed the token
