+++ src/admin.py
@@ -18,5 +18,6 @@ def setup_admin(app):
     
     
     
+    
     # You can duplicate that line to add mew models
     # admin.add_view(ModelView(YourModelName, db.session))
